Component TopTabBarNavigator renamed to PeopleScreenNavigator and moved into people folder (+ all related files)
