+++ mobile/src/reducers/navigation.reducer.ts
@@ -1,6 +1,6 @@
 import { RootNavigator } from '../tabbar/tab-navigator';
 import { NavigationAction, NavigationState } from 'react-navigation';
-import { TopTabBarNavigator } from '../toptabbar/top-tab-bar-navigator';
+import { PeopleScreenNavigator } from '../people/navigator/people-screen-navigator';
 
 export const navigationReducer = (state: NavigationState, action: NavigationAction) => {
     const nextState = RootNavigator.router.getStateForAction(action, state);
@@ -8,6 +8,6 @@ export const navigationReducer = (state: NavigationState, action: NavigationActi
 };
 
 export const peopleNavigationReducer = (state: NavigationState, action: NavigationAction) => {
-    const nextState = TopTabBarNavigator.router.getStateForAction(action, state);
+    const nextState = PeopleScreenNavigator.router.getStateForAction(action, state);
     return nextState || state;
 };
